+++ model-implementation/src/main/java/org/verapdf/model/impl/pb/xmp/schemas/PBPDFAIdSchema.java
@@ -32,7 +32,7 @@ public class PBPDFAIdSchema extends PBXMPPredefinedSchema implements PDFAIdSchem
      */
     @Override
     public Long getpart() {
-        return ((PDFAIdentificationSchema) getXmpSchema()).getPart() == null ? null : Long.valueOf(((PDFAIdentificationSchema) getXmpSchema()).getPart());
+        return ((PDFAIdentificationSchema) getXmpSchema()).getPart() == null ? null : Long.valueOf(((PDFAIdentificationSchema) getXmpSchema()).getPart().longValue());
     }
 
     /**
